Move extension factory to a separate package
