+++ packages/zeplin-css-extension/src/index.js
@@ -1,4 +1,4 @@
-import { extensionFactory } from "zeplin-extension-style-kit/factory";
+import { createExtension } from "base-extension";
 
 import Generator from "./generator";
 import { COPYRIGHT } from "./constants";
@@ -7,7 +7,7 @@ const exportPrefix = `${COPYRIGHT}\n\n`;
 
 const INDENTATION =  "  ";
 
-export default extensionFactory({
+export default createExtension({
     language: "css",
     Generator,
     colorsOptions: {
