+++ packages/zeplin-stylus-extension/package.json
@@ -11,6 +11,7 @@
     "publish": "zem publish"
   },
   "dependencies": {
+    "base-extension": "^1.0.0",
     "zem": "^1.0.6",
     "zeplin-extension-style-kit": "^2.1.2"
   },
